Added WallpaperDetailsScreen and related components.
This commit introduces the `WallpaperDetailsScreen` composable, which displays detailed information about a selected wallpaper. It includes the following components and changes:

- **UI Components:**
    - `WallpaperDetailsScreen`: The main composable for the details screen.
    - `WallpaperDetailsScreenContainer`: Holds the main content of the details screen in a `LazyColumn`.
    - `WallpaperDetailsScreenTopBar`: Displays the screen title and navigation actions.
    - `WallpaperDetailsScreenBottomBar`: Contains actions like download.
    - `WallpaperDetailsScreenWallpaperItem`: Displays the wallpaper image and views count.
    - `WallpaperDetailsScreenUploaderItem`: Shows uploader information and creation date.
    - `WallpaperDetailsScreenTagsItem`: Displays wallpaper tags using `TagLayout` and `TagItem`.
    - `WallpaperDetailsScreenOtherItem`: Shows additional wallpaper details like resolution, size, type, and aspect ratio.

- **ViewModel and State:**
    - `WallpaperDetailsScreenViewModel`: Handles the logic for fetching and displaying wallpaper details.
    - `WallpaperDetailsScreenState`: Represents the UI state of the details screen.
    - `WallpaperDetailsScreenUIAction`: Defines user interactions on the details screen.

- **Navigation:**
    - Integrated `WallpaperDetailsScreen` into the `wallpapersNavGraph`.
    - Updated `WallpapersNavGraphTransitions` to handle transitions to and from the details screen, considering layout direction.
    - `NavScreen` now implements `Parcelable`.

- **Data Models and Mappers:**
    - Updated `WallpaperEntity` and `WallpaperResponseData` to use `Int` for `views` instead of `Long`.
    - `WallpaperModel`, `WallpaperThumbnailsModel`, and `WallpaperUploaderModel` are now `@Stable`.
    - Enhanced `WallpaperEntityToModelMappers` to format file size, MIME type, and calculate aspect ratio.

- **Utilities:**
    - Added `DateTimeUtils` for formatting dates.
    - Added `DataUtils` for converting bytes to a human-readable format.
    - Added `MimeTypeUtils` for converting MIME types to readable strings.
    - Added `NumberUtils` for formatting numbers (e.g., view counts).
    - Added `ProportionsUtils` for calculating aspect ratios.
    - Added `LazyListStateExt` with an `isScrolledToTheEnd()` extension.

- **Core UI:**
    - Introduced `TagLayout` and `TagItem` composables for displaying tags.

- **Other Changes:**
    - Enabled navigation to `WallpaperDetailsScreen` from `WallpapersHomeScreenViewModel`.
    - Removed unused imports in `WallpaperResponseToEntityMappers`.
    - Updated string resources.
    - Added "mgtpe" to the project dictionary.

+++ app/src/main/java/com/stoyanvuchev/stylepaper/feature_wallpapers/data/local/entity/WallpaperEntity.kt
@@ -47,6 +47,6 @@ data class WallpaperEntity(
     val thumbs: WallpaperResponseThumbnails,
     val uploader: WallpaperResponseUploader,
     val url: String,
-    val views: Long
+    val views: Int
 
 )
