+++ app/src/main/java/com/stoyanvuchev/stylepaper/feature_wallpapers/data/remote/response/WallpaperResponseData.kt
@@ -71,6 +71,6 @@ data class WallpaperResponseData(
     val url: String,
 
     @SerializedName("views")
-    val views: Long
+    val views: Int
 
 )
